♻️ refactor: tag page migration
